now marking skipped steps with yellow
